% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsem-package.R, R/ctsem.R
\docType{package}
\name{ctsemOMX}
\alias{ctsemOMX}
\alias{ctsemOMX-package}
\title{ctsemOMX}
\description{
ctsem is an R package for continuous time structural equation modelling of panel (N > 1) 
and time series (N = 1) data, using either a frequentist or Bayesian approach, or middle
ground forms like maximum a posteriori. This ctsemOMX addition includes the original OpenMx based functions
which have been split off from the main package.

ctsem is an R package for continuous time structural equation modelling of panel (N > 1) 
and time series (N = 1) data, using either a frequentist or Bayesian approach, or middle
ground forms like maximum a posteriori. This ctsemOMX addition includes the original OpenMx based functions
which have been split off from the main package.
}
\details{
The general workflow begins by specifying a model using the \code{\link{ctModel}} function, 
in which the \code{type} of model is also specified. Then the model is fit to data using 
either \code{\link{ctFit}} if the original 'omx' (OpenMx, SEM, max likelihood) model is specified.
The omx forms are no longer in 
development and for most purposes, the newer stan based forms (contained in the base ctsem package)
are more robust and flexible.
For citation info, please run \code{citation('ctsem')} .

The general workflow begins by specifying a model using the \code{\link{ctModel}} function, 
in which the \code{type} of model is also specified. Then the model is fit to data using 
either \code{\link{ctFit}} if the original 'omx' (OpenMx, SEM, max likelihood) model is specified.
The omx forms are no longer in 
development and for most purposes, the newer stan based forms (contained in the base ctsem package)
are more robust and flexible.
For citation info, please run \code{citation('ctsem')} .
}
\references{
https://www.jstatsoft.org/article/view/v077i05

Driver, C. C., & Voelkle, M. C. (2018). Hierarchical Bayesian continuous time dynamic modeling. 
Psychological Methods. Advance online publication.http://dx.doi.org/10.1037/met0000168

https://www.jstatsoft.org/article/view/v077i05

Driver, C. C., & Voelkle, M. C. (2018). Hierarchical Bayesian continuous time dynamic modeling. 
Psychological Methods. Advance online publication.http://dx.doi.org/10.1037/met0000168
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdriveraus/ctsemOMX}
}

}
\author{
\strong{Maintainer}: Charles Driver \email{charles.driver@ibe.uzh.ch} [copyright holder]

Authors:
\itemize{
  \item Manuel Voelkle [copyright holder]
  \item Han Oud [copyright holder]
}

}
\keyword{internal}
